requirements.txt
# Story Theory Benchmark Dependencies

# API and HTTP
openai>=1.0.0
httpx>=0.25.0

# Environment and Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Data Validation
jsonschema>=4.20.0
pydantic>=2.5.0

# Statistical Analysis
scipy>=1.10.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Data Processing
pandas>=2.0.0

# Progress and CLI
tqdm>=4.66.0
rich>=13.0.0

# Testing
pytest>=7.4.0

# HuggingFace Integration (for publication)
datasets>=2.15.0
huggingface-hub>=0.19.0
